[REFACTOR] Split out the date time formatter to separate file
